pyproject.toml
[tool.poetry]
name = "kmibot"
version = "0.1.0"
description = ""
authors = ["Hiroaki ITO <mail@isudzumi.com>"]

[tool.poetry.dependencies]
python = "^3.9"
transformers = "^4.10.0"
line-bot-sdk = "^1.20.0"
sentencepiece = "^0.1.96"
torch = "^1.9.0"

[tool.poetry.dev-dependencies]
boto3 = "^1.18.36"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
